--- a/Project.toml
+++ b/Project.toml
@@ -5,8 +5,6 @@ version = "0.2.2"
 
 [deps]
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [weakdeps]
@@ -21,10 +19,8 @@ JacobiEllipticZygoteExt = "Zygote"
 
 [compat]
 DocStringExtensions = "0.9"
-Documenter = "1.1"
 Enzyme = "0.11, 0.12"
-Zygote = "0.6"
 ForwardDiff = "0.10"
-Setfield = "1.1"
 StaticArrays = "1.6"
+Zygote = "0.6"
 julia = "1.8"