I noticed that JacobiElliptic depended on Documenter. Typically, packages don't directly depend on Documenter since it is a heavy dependency. I've removed that since it doesn't appear it was used for anything.

Additionally, I noticed that Setfield was solely being used for changing a SVector in a couple of spots in the code. I've removed the need for that by switching to an MVector and throwing in a couple of inbounds. In newer Julia versions, if the MVector doesn't escape the function (i.e. not returned) then there typically is no allocation. I've checked this locally and don't see any allocations. Plus, at least locally, removing the `@set` seems to speed up the code slightly. 

After this PR

```julia
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  25.163 ns … 39.163 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     25.326 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   25.443 ns ±  0.508 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▅█▆  ▁▄▃         ▂                ▁                       ▁
  ▅▇████▆███▆▄▃▁▃▁▃▁▇█▇▁▁▆▇▄▁▁▁▄▃▆▆▆▇██▇▇▆▆▄▅▅▅▅▆▅▅▅▄▄▅█▆▃▁▄▄ █
  25.2 ns      Histogram: log(frequency) by time      27.7 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

Before this PR

```julia
BenchmarkTools.Trial: 10000 samples with 995 evaluations.
 Range (min … max):  26.741 ns … 45.222 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     26.808 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.932 ns ±  0.561 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅█▄ ▄▅            ▁                                         ▁
  ███████▅▁▁▃▁▁▁▁▁▃██▅▁▄▅▁▁▃▁▁▄▅▇▇▇▇▆█▇▅▆▅▄▅▄▆▅▄▅▃▅▅▆▃▅▅█▆▅▅▄ █
  26.7 ns      Histogram: log(frequency) by time      29.3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```